Fix Windows worker crashing
**Issue:**
https://github.com/rancher/rancher/issues/17499
